+++ assets/index.js
@@ -1,4 +1,4 @@
-const { change_page ,left_input,right_input} = require("./navigation")
+const {left_input,right_input} = require("./navigation")
 const {changeTarget,findBySubString,loadAllSuggestions,postSuggestion,loadPostsFromCategory,destroyPosts} = require("./suggestions")
 
 window.addEventListener("resize", () => width = window.innerWidth)
@@ -20,6 +20,6 @@ if(window.location.href.includes("suggestions.html")){
         loadPostsFromCategory();
     }))
     document.getElementById("search-form").addEventListener("submit",findBySubString)
-    window.addEventListener("load", async ()=> loadAllSuggestions())
+    window.addEventListener("DOMContentLoaded", async ()=> loadAllSuggestions())
     document.getElementById("post").addEventListener("submit",postSuggestion)
 }
